+++ src/net/psj/Walls/WallFan.java
@@ -5,6 +5,6 @@ import net.psj.Interface.MenuData;
 public class WallFan extends Wall {
 
 	public WallFan() {
-		super("FAN", new float[] { 0.0f, 0.0f, 1.0f }, 2, MenuData.MENU_WALLS);
+		super("FAN", new float[] { 0.0f, 0.0f, 1.0f }, (short) 2, MenuData.MENU_WALLS);
 	}
 }
